Fixed pernicious bug related to memoization.
The solution of an env using smooth_rare was inexplicably switched to
accelshock, but only very occasionally, in the first observed case not
happening for 700+ training steps.
This was a result of not copying the output of "get_full", which returns
the data itself so that one may overwrite it, if desired.
It was specific to accelshock because only accelshock resets the grid in
place instead of declared a new ndarray.
That said, there were probably other side-effects I didn't notice. Good
thing I haven't been using this for too long.