Param sweep can handle dependent parameters.
Instead of exclusively independent parameters. This allows for cases
when you want a value for one parameter to also have another parameter
set.
So we can have
--init-type sine
and
--init-type schedule --eval-env custom
handled as different init-type configurations.

Not perfect, we still need to be able to handle e.g.
--init-type sine;--init-type random --seed 1;--init-type random --seed 2
in a way that puts the seed in the name of the log directory.