New structure: algorithm = (algo + model).
util/run.py moved to algos/run.py.

The learn method in run.py now takes an algo; the algo gets an env and
can do whatever with it.
The algos should be designed to function around an arbitrary model.
This way, the algo is orthogonal to the environment, model, and
the miscellaneous training management in learn().