Added action/state adjustments into new structure.
Moved into EMI, which makes sense. Removed some things like the ability
to change the squash function on the policy network.
The functions are still defined in run_train, but are then passed to the
EMI to use as part of its interface.

Should be good to go, just need to test.

Also changed .gitignore to hide all of the logs instead of just the ones
in the log and test.