Fixed bug prevent SAC wrapper from working.
The bug was actually in UnbatchedEnvPL, an instance of which is passed
to SACModel. The high of the action space was being set to the same as
the low of the action space. This meant that the actions from the model
were being scaled by (low-low) instead of (high-low), so the actions
were always 0.0.

I'm still not convinced that it's working, though. Need to poke around
more.