Parameterized rewards.
Arguments can now be passed to run_train and run_test to specify
different types of rewards.
I hesitated to make this change because it fundamentally changes the
environment, so really should be specified by changing the environment.
However, this is useful for using different reward types in a paramter
sweep.