Implemented more of the network in TF.
There was more to do than I thought. The built-in RNN does not handle
this case, so I implemented it myself. Also implemented the reward
function, which is involved considering it needs to include computing
the next state with WENO. That feels wrong, but I don't know any other
way to do this.