Resolved some issues in network implementation.
Namely these 2:
The policy needs to be a subclass of Layer instead of just a function
because we need to use it in multiple places, that is, repeatedly inside
of the global network, and on its own during inference. I could have
used reuse=True isntead, but I prefer this abstraction.
Inside the recurrent cell, the policy expects a batch, but of individual
states, not every state at once, so the first batch and location axes
of the state need to be combined before passing it to the policy.