Resolved discrepancy with ghost cells.
It wasn't clear when to fill them - I've put them in the prep_state
function that converts real state to rl_state. This might not be the
ideal choice, though - the reward calculation assumes all errors past
the boundary are zero, but if it didn't it would need access to the
ghost cells that don't exist at that point in the calculation.