Moved TF functions into env file.
The 3 functions hardcoded into the network rely on information available
only to the environment (such as private fields like ng and nx), so they
were moved into burgers_env.py instead.
Only WENOBurgersEnv versions are implemented, but the code is written
to handle SplitFluxBurgersEnv versions instead, were they to be
implemented in the future.