Fixed redundant line (thanks, Yiwei).
The predict() method in SACWrapper had 2 lines like
action = sac.predict()
action = sac.policy_tf.step()
This should not be any different because the 2nd line, the correct
version, overwrites the first, but is redundant and wastes CPU.
The bug comes from my original implementation of the function. predict()
in SACWrapper was supposed to simply forward the call to sac.predict(),
but I realized that wouldn't work because of the way I had things set
up, so the predict() needs to handle both training and testing.
Compare the predict() function and the relevant section of the learn()
function in the original stablebaselines/sac/sac.py file.