Added util/function_dict.py with np/tf functions.
A lookup table of functions associated with certain string keys. This
is something I wanted to do anyway - strings are easier to serialize
than functions, and I had copied and pasted these functions so they
could be used in multiple places.
This version has the lookup table for Numpy functions, but also has a
Tensorflow function for each key. This is necessary for implementing the
action/observation adjustments inside the model.