Model integrates action/obs scaling functions.
The model keeps a clean copy of the PolicyNet for use in predict()
calls, but wraps the PolicyNet in a FunctionWrapper before passing it to
the recurrent network. FunctionWrapper is a new class that applies
functions to its input and output, such as those returned from
tensorflow_fn in util.function_dict.