Possible memory leak fix.
Part of checking for NaNs inadvertently added a new piece to the graph
at every iteration. Small, but enough to be a problem in the long run.
The NaN check is now declared when the rest of the graph is created, and
graph.finalize() is called to detect such things in the future.