Fixed(?) L2 error during training for full model.
Previously it was not recording it all, because it did not use the
actual environment passed to it.
Now the EMI must return the L2 error in addition to the reward.
I'm not sure if it's quite right, however, as it seems much too high
relative to the eval environments.