full model construction is lazy.
That is, it is not constructed until required by e.g. calling train().
This is useful for testing as it allows us to load only the policy part
of the model without needing specific information like the number of
cells.

Relatedly, --convergence-plot works again for run_test.py.

Sorry about the hassle this will probably be to merge with.