Fixed the leak. Caused by saving actions.
The reset method didn't reset the action history to [].
Strange that this didn't become an issue until now; I guess the leak is
small enough to not be a problem for a long time, but becomes a much
bigger issue if we store 10 episodes worth of actions every episode.

The reset() method in AbstractBurgersEnv was changed from
NotImplementedError to doing some of the standard reset behavior,
including resetting the acting history, with the expectation that
subclasses call super().reset() before return the initial state.