Fixed loss function reduction dimensions.
To be interpretable as the average return, it needs to sum over axis 0,
not axis 2.

Also: fix to how the policy parameters are declared. No change in
behavior, as the previous version happened to work despite being
incorrect. (But didn't work when I was debugging a thing.)

Also: added a comment as a reminder to myself and anyone else looking at
it as to why the policy accessed by the predict() method in the model
does not include the normalizing functions.