Linked dt to dx.
In other words, made the default timestep length dependent on the
explicitly specified nx, or the default nx dependent on the explicitly
specified timestep length.
Also added the related --time-max parameter, which overrides the
ep-length parameter and is useful to maintain a constant total time
when the timestep length varies.
There's a bug in this behavior - when running a test with an existing
model, both the nx and timestep length from the model's metadata are
considered explicit parameters, so using e.g. a different nx with the
model will use the timestep length given in the model's metadata,
instead of one computed based on the new nx.