Created Grid2d class.
This involved refactoring the GridBase and Grid1d classes into
AbstractGrid, which controls the idea of a cell-centered grid,
GridBase, which controls the actual representation of the grid,
and Grid1d, which controls how the grid is initialized.

A new class AxisSlice was add to util.misc. AxisSlice is handy for
making numpy code on variable dimensions more readable.