Added "1d" family of initial conditions to Grid2d.
These can be expressed by e.g. '1d-sine-x' where a trailing -x or -y
is the dimension that will vary as per the equivalent initial condition,
and the remaining dimension will be the same across the space.

Passing simply '1d' and them giving parameters 'type': 'sine' and
'axis':'x' also works.

This involves allowing for multi-dimensional boundary conditions. If the
boundary condition is a scalar, it will be expanded to every dimension,
otherwise each element of the boundary condition applies to a different
dimension. Boundary conditions are evaluated first-to-last.