Adjusted grid so 1D is not an exceptional case.
It assumes things are tuples/lists, and converts things if they're not.
1-dimensional stuff is still accessible via the original names e.g. nx,
ng, dx, etc.
The exception to this is grid.boundary, which may be either a tuple or a
string that applies to every dimension. This is partially for
compatability, and partially for convenience in the reset() function
of subclasses.