Created AbstractScalarEnv.
A high level abstract environment to handle scalar equations in
arbitrary dimensions.
Later we should expand it to vector equations - not sure what will be
needed for that.

Renamed the file from 2d_burgers_env.py to abstract_scalar_env.py, given
the switch to the more abstract class first.