Added 'combo' 2D initial condition.
This combines a 1D initial condition on the x axis and a 1D initial
condition on the y axis additively.

--init-type combo will default to smooth_sine on both dimensions.
--init-type combo-TYPE will use TYPE on both dimensions.
--init-typei combo-TYPE1-TYPE2 will use TYPE1 on the x axis and TYPE2 on
the y axis.

Interestingly, some combinations result in an error between the standard
WENO agent and the WENO solution. This should be impossible as they're
using the same code - what's happening here?