Fix for 1D convergence plots.
They always use the original number of cells instead of varying.
The error was caused by changing 'nx' to 'num_cells'. The arguments for
the convergence plot environments still change nx, and then nx is
printed later so it seems as if it is working. However, nx is not used
by the environment; instead it uses the existing num_cells.