Figured out how to determine explicit args.
Involves using a sentinel object and the existing behavior of the
namespace argument in ArgParser.parse_args().

Also added serialize() and load_from_dict() to parameter_dict.py.
Converting a serialized string into a dict can be done by the Yaml parser, though we
could write a simple parser if we wanted to avoid the external requirement.