Conditional sub-parsing.
That is, the parsers for model and environment parameters may now be
built to depend on main parameters.
Currently, the model parser now only shows arguments relevant to the
--model parameter parsed by the main parser.
This includes when loading from a file; a lot of repeated parsing is
necessary to do this, but that's not so big a cost.