Fix time bug in solution update.
solution.update(dt, time) should be called with the time AFTER the
timestep, not before. That is, the solution should update to the state
at the time given.
This doesn't make a difference for the weno solutions, which only use
the dt argument anyway. However, analytical solutions do make use of
time, so they were consistently off by one timestep.

Also, make convergence plots more like the ones in the hydrodynamics
book. Some related bugfixes as well - gaussian should be outflow, not
periodic.