Implement SSP RK3.
The syntax for using an RK method is now e.g. --rk rk4 or --rk ssp_rk3,
as opposed to before when it was simply --rk4.

The flow needed to change slightly in step execution. Previously,
env._finish_step() expected a step and a timestep; now it accepts the
next state and a timestep.
A small change, but needed for the implementation of SSP RK3.

There is now an enum in envs/weno_solutions.py, RKMethod, which
enumerates the available RK method (Euler, RK4, and SSP_RK3).