Corrected final dynamic timestep.
Previously with dynamic timesteps, the episode was considered finished
if the time was greater than the end time, but this meant that you could
go slightly over that time.
Now, the timestep is cropped if it would go over the end time.
(Note that this is not implemented in global backprop, but it should be
less important during training.)