Eval envs during training should be standardized.
Was previously ep_length * 2, but they should just use 500 (or whatever
the default is for that initial condition). Similarly for other
defaults. The exception is order, which does need to be the same for
training and testing environments.
Also added a prefix to the print messages in
set_contingent_env_defaults() for clarity when setting the parameters of
multiple environments this way.