Added half_evolution_plot.py.
This script creates an evolution plot that only has lines for halfway
through an episode and at the very end.
Intended specifically for the paper and contains a complex legend that
only fits on smooth_sine, with the intent that other such plots are
nearby and do not need a legend.