scripts/reeval.py is now thread safe.
Or rather, process safe. This handles the potential risk of writing to
the progress.csv file simultaneously by using util.filelock.

Something that's been bothering me for a while.

Also fixed a bug in scripts/reeval.py related to merging the original
experiment file with an existing file in the target directory. The
problem was that Pandas sometimes rounds floats for some reason.