TF implementation of weno_interpolation.
tf_weno_interpolation acts the same as weno_interpolation except with TF
tensors.
I ran into a known bug with tf.linalg.diag_part which is supposed to use
matrix_diag_part_v2. It's fixed in TF2, but not TF1.15, and they aren't
going to fix it at this point.